requirements.txt
blinker==1.8.2
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
cryptography==43.0.3
dnspython==2.7.0
Flask==3.0.3
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
Naked==0.1.32
pycparser==2.22
pycryptodome==3.21.0
pycryptodomex==3.21.0
pymongo==4.10.1
PyYAML==6.0.2
requests==2.32.3
shellescape==3.8.1
urllib3==2.2.3
Werkzeug==3.1.2
